backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
click==8.3.3
colorama==0.4.6
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
librt==0.9.0
Mako==1.3.11
MarkupSafe==3.0.3
mypy==1.20.2
mypy_extensions==1.1.0
packaging==26.2
pathspec==1.1.1
pluggy==1.6.0
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
pytest==9.0.3
pytest-asyncio==1.3.0
python-dotenv==1.2.2
PyYAML==6.0.3
ruff==0.15.12
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
watchfiles==1.1.1
websockets==16.0
celery[redis]>=5.3.0
python-magic>=0.4.27
python-multipart>=0.0.9
redis[asyncio]>=5.0.0
